Initial commit.
This commit uses the WordPress Theme development starter called `_s`
("Underscores"). It is the stock Underscores theme with the name itself
(`_s`) changed to The Lark's needs. No functional changes have been
made, though, and so this commit represents the initial state of the
Underscores codebase as delivered by the WordPress Core team.